refactor(AdminSchoolDetail): データ取得ロジックをhooksに切り出し
